test(sys): add stmt test case
